\name{FlowRepositoryR-package}
\alias{FlowRepositoryR-package}
\alias{FlowRepositoryR}
\docType{package}
\title{Programming interface to access FlowRepository.org}

\description{
    This package includes functions to locate and review FlowRepository datasets
    and conveniently download related data to the local file system.
}

\references{
    Josef Spidlen. FlowRepository API and Resources for Developers.\cr
    \url{http://flowrepository.org/developers}\cr\cr

    Josef Spidlen, Wayne Moore, David Parks, Michael Goldberg, Chris Bray, 
    Pierre Bierre, Peter Gorombey, Bill Hyun, Mark Hubbard, Simon Lange, 
    Ray Lefebvre, Robert Leif, David Novo, Leo Ostruszka, Adam Treister, 
    James Wood, Robert F. Murphy, Mario Roederer, Damir Sudar, Robert Zigon, 
    and Ryan R. Brinkman.\cr
    Data File Standard for Flow Cytometry, Version FCS 3.1\cr
    Cytometry A. 2010 Jan; 77(1): 97100. doi: 10.1002/cyto.a.20825\cr
    \url{http://tinyurl.com/fcs3-1-pdf}
}

\author{
    Author and maintainer: Josef Spidlen <jspidlen@bccrc.ca>
}

\keyword{package}
